Move old scripts and functions to obsolete dir
