+++ eniric/__init__.py
@@ -2,7 +2,6 @@ __version__ = "0.1"
 __all__ = [
     "atmosphere",
     "IOmodule",
-    "nIRanalysis",
     "plotting_functions",
     "Qcalculator",
     "resample",
