+++ eniric/atmosphere.py
@@ -345,104 +345,6 @@ class Atmosphere(object):
         )
 
 
-def barycenter_shift(
-    wav_atm: ndarray,
-    mask_atm: ndarray,
-    rv_offset: float = 0.0,
-    consecutive_test: bool = True,
-) -> ndarray:
-    """Calculating impact of Barycentric movement on mask...
-
-    Extends the masked region to +-30 km/s due to the barycentric motion of the earth.
-    """
-    # Mask values to the left and right side of mask_atm.
-    # To avoid indexing errors have padded with first and last values.
-    mask_iminus1 = np.concatenate(
-        ([mask_atm[0], mask_atm[0]], mask_atm[:-2])
-    )  # padding with first value
-    mask_iplus1 = np.concatenate(
-        (mask_atm[2:], [mask_atm[-1], mask_atm[-1]])
-    )  # padding with last value
-
-    pixels_total = len(mask_atm)
-    masked_start = pixels_total - np.sum(mask_atm)
-
-    barycenter_rv = 3e4  # 30 km/s in m/s
-    offset_rv = rv_offset * 1.0e3  # Convert to m/s
-
-    # Doppler shift  applied to the vectors
-    delta_lambdas = wav_atm * barycenter_rv / const.c.value
-    offset_lambdas = wav_atm * offset_rv / const.c.value  # offset lambda
-
-    # Doppler shift limits of each pixel
-    wav_lower_barys = wav_atm + offset_lambdas - delta_lambdas
-    wav_upper_barys = wav_atm + offset_lambdas + delta_lambdas
-
-    mask_atm_30kms = np.empty_like(mask_atm, dtype=bool)
-
-    for i, (wav_value, mask_val) in enumerate(zip(wav_atm, mask_atm)):
-        """If there are 3 consecutive zeros within +/-30km/s then make the value 0."""
-
-        # Offset_RV is the offset applied for the star RV.
-        if (
-            (mask_val is False)
-            and (mask_iminus1[i] is False)
-            and (mask_iplus1[i] is False)
-            and (rv_offset == 0)
-        ):  # If the mask is false and the offset is equal to zero
-            """If the value and its 2 neighbours are already zero don't do the barycenter shifts"""
-            mask_atm_30kms[i] = False
-        else:
-            # np.searchsorted is faster then the boolean masking wavelength range
-            # It returns index locations to place the min/max doppler-shifted values
-            slice_limits = np.searchsorted(
-                wav_atm, [wav_lower_barys[i], wav_upper_barys[i]]
-            )
-            slice_limits = [
-                index if (index < len(wav_atm)) else len(wav_atm) - 1
-                for index in slice_limits
-            ]  # Fix searchsorted index
-
-            mask_atm_slice = mask_atm[slice_limits[0] : slice_limits[1]]
-
-            mask_atm_slice = np.asarray(
-                mask_atm_slice, dtype=bool
-            )  # Insure boolean dtype
-
-            if consecutive_test:
-                # Make mask value False if there are 3 or more consecutive zeros in slice.
-                len_consec_zeros = consecutive_truths(~mask_atm_slice)
-                if np.all(
-                    ~mask_atm_slice
-                ):  # All pixels of slice is zeros (shouldn't get here)
-                    mask_atm_30kms[i] = False
-                elif np.max(len_consec_zeros) >= 3:
-                    mask_atm_30kms[i] = False
-                else:
-                    mask_atm_30kms[i] = True
-                    if np.sum(~mask_atm_slice) > 3:
-                        print(
-                            "There were {0}/{1} zeros in this barycentric shift but None were 3 consecutive!".format(
-                                np.sum(~mask_atm_slice), len(mask_atm_slice)
-                            )
-                        )
-            else:
-                mask_atm_30kms[i] = np.product(mask_atm_slice)
-    masked_end = pixels_total - np.sum(mask_atm_30kms)
-    print(
-        (
-            "New Barycentric impact affects the number of masked pixels by {0:04.1%} due to the atmospheric"
-            " spectrum"
-        ).format((masked_end - masked_start) / pixels_total)
-    )
-    print(
-        ("Masked Pixels start = {1}, masked_pixel_end = {0}, Total = {2}").format(
-            masked_end, masked_start, pixels_total
-        )
-    )
-    return mask_atm_30kms
-
-
 def consecutive_truths(condition: ndarray) -> ndarray:
     """Length of consecutive true values in an bool array.
